be nearly finished, done error handling logic, todo: more detailed errors in catch blocks in abl

+++ abl/list/remove-user-abl.js
@@ -3,6 +3,8 @@ const ListDao = require('../../dao/list-dao')
 const UserDao = require('../../dao/user-dao')
 
 const isOwner = require('../../abl/utils/user-validation')
+const { ABL_ERRORS } = require('../../errors/abl')
+const { DaoError } = require('../../errors/dao')
 
 const listDao = new ListDao()
 const userDao = new UserDao()
@@ -17,34 +19,43 @@ const paramsSchema = {
 }
 
 const RemoveShoppingListUserAbl = async (req, res) => {
-    const userId = req.user.id
-    const { id } = req.params
+    try {
 
-    if (!await isOwner(userId, id)) {
-        res.status(403).json({ message: 'Forbidden' })
-        return
+        const userId = req.user.id
+        const { id } = req.params
+
+        if (!await isOwner(userId, id)) {
+            return res.status(403).json({ code: ABL_ERRORS.forbidden, message: 'You are not allowed to call this feature' })
+        }
+
+        const ajv = new Ajv()
+
+        const validParams = ajv.validate(paramsSchema, req.params)
+        if (!validParams) {
+            return res.status(400).json({ code: ABL_ERRORS.invalidParams, message: ajv.errors })
+        }
+
+        const removedUserId = req.params.userId
+        if (removedUserId === userId) {
+            return res.status(400).json({ code: ABL_ERRORS.invalidParams, message: 'Owner cannot leave the list' })
+        }
+
+        let result = await listDao.removeShoppingListUser(id, removedUserId)
+
+        const members = await userDao.getUsersById(result.members)
+        const owner = members.find(member => member.id === result.owner)
+
+        res.json({
+            ...result,
+            members,
+            owner
+        })
+    } catch (err) {
+        if (err instanceof DaoError) {
+            return res.status(400).json({ code: err.code, message: err.message })
+        }
+        res.status(500).json({ code: ABL_ERRORS.unknown, message: err.message })
     }
-
-    const ajv = new Ajv()
-
-    const validParams = ajv.validate(paramsSchema, req.params)
-    if (!validParams) {
-        res.status(400).json(ajv.errors)
-        return
-    }
-
-    const { removedUserId } = req.params
-
-    let result = await listDao.removeShoppingListUser(id, removedUserId)
-
-    const members = await userDao.getUsersById(result.members)
-    const owner = members.find(member => member.id === result.owner)
-
-    res.json({
-        ...result,
-        members,
-        owner
-    })
 }
 
 module.exports = RemoveShoppingListUserAbl
